Begin draft of LRUG "statement of intent"
This roughly summarises the conversation that we (@aanand, @h-lame,
@tomstuart) had in the pub last night. We said we'd like to create
a unified document that explains the purpose of LRUG and outlines
the community's expectations of the conduct of its members, as well
as providing a basic incident handling mechanism for anyone who has
been made to feel unwelcome or unsafe.

There's no real format or structure yet, but this draft will serve
as a reminder of the main points we'd like to cover.