Don't duplicate #content
We use #content for the main wrapper div and as the title of a section, and so the ToC links took you to the wrong place.  The quick fix is to use a thesaurus and replace the "Content" heading with "Subject Matter".  Revisit when we come up with nicer css.