Add an 'About this document' section.
This commit adds an 'About this document' section at the foot of the
document. It is intended to address some of the points raised in

It adds a short paragraph on the origin of the document, and then
outlines how subsequent, and future changes to it can be made. I've
hand-waved a bit on how proposed changes are incorporated into the
document, but based it on the closed issues in the github repo, that is,
the discussion around any changes is open and in the public, and when
consensus is reached the changes are merged in.

On the issue of volunteering, to address primarily how named people in
the document are 'appointed', I've added a short paragraph on
volunteering, which encourages the reader to contact @hlame in the first
instance.