Retire the 20x20 format for lightning talks
Historically we’ve done February lightning talks in the 20x20 format,
but there was concern [1] that it might be offputting for new speakers,
so we’ve now dropped it in favour of a simple 10-minute time limit. This
is how the February 2015 meeting worked [2] and will probably remain the
case for future Februaries.

[1] http://lists.lrug.org/pipermail/chat-lrug.org/2015-January/010669.html
[2] http://lrug.org/meetings/2015/february/